feat(settings): editable practice jurisdictions; enforce update permission
Adds an `OrganizationJurisdictionsCard` to the org members settings
page so jurisdictions can be edited after onboarding (the
`jurisdictionSaveFailed` toast already promised this). Picker UI
extracted into a shared `JurisdictionPicker` reused by the
onboarding step and the new settings card.

Wires the missing `permissions` option on the
`POST /organization-settings/practice-jurisdictions` route so
`organizationSettings:update` is actually enforced — without it,
any authenticated org member could update org-wide jurisdiction
context. (P2 from Codex review on PR #237.)
