+++ apps/web/src/i18n/langs/cs.json
@@ -1399,6 +1399,7 @@
       "matterNumberingDescription": "Nastavte, jak se generují referenční čísla nových spisů",
       "membersDescription": "Spravujte, kdo má přístup k této organizaci, a čekající pozvánky",
       "pendingInvitations": "Čekající pozvánky",
+      "practiceJurisdictions": "Jurisdikce praxe",
       "profile": "Profil",
       "profileDescription": "Upravte název organizace",
       "renameAction": "Rename",
