+++ apps/web/src/i18n/langs/en.json
@@ -1399,6 +1399,7 @@
       "matterNumberingDescription": "Configure how new matter reference numbers are generated",
       "membersDescription": "Manage who has access to this organization and any pending invitations",
       "pendingInvitations": "Pending invitations",
+      "practiceJurisdictions": "Practice jurisdictions",
       "profile": "Profile",
       "profileDescription": "Update organization name",
       "renameAction": "Rename",
