+++ apps/web/src/i18n/langs/pl.json
@@ -1399,6 +1399,7 @@
       "matterNumberingDescription": "Configure how new matter reference numbers are generated",
       "membersDescription": "Manage who has access to this organization",
       "pendingInvitations": "Pending invitations",
+      "practiceJurisdictions": "Jurysdykcje praktyki",
       "profile": "Profile",
       "profileDescription": "Update name and slug",
       "renameAction": "Rename",
