+++ apps/web/src/i18n/langs/sk.json
@@ -1399,6 +1399,7 @@
       "matterNumberingDescription": "Nakonfigurujte, ako sa generujú referenčné čísla nových spisov",
       "membersDescription": "Spravujte, kto má prístup k tejto organizácii a všetky čakajúce pozvánky",
       "pendingInvitations": "Čakajúce pozvánky",
+      "practiceJurisdictions": "Jurisdikcie praxe",
       "profile": "Profil",
       "profileDescription": "Upraviť názov organizácie",
       "renameAction": "Premenovať",
